Refactor: Reorganize the project structure and extract the app to the homepage.
